backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
mysqlclient>=2.1.0
requests>=2.31.0
PyJWT>=2.8.0
python-decouple>=3.8